AnswerRepository changed

+++ StudentFeedback/src/main/java/at/htl/survey/controller/AnswerRepository.java
@@ -65,13 +65,13 @@ public class AnswerRepository implements Persistent<Answer>{
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(long id) {
 
         try (Connection connection = dataSource.getConnection()) {
             String sql = "DELETE FROM answer  WHERE a_id=?";
 
             PreparedStatement statement = connection.prepareStatement(sql);
-            statement.setInt(1, id);
+            statement.setLong(1, id);
 
             if (statement.executeUpdate() == 0) {
                 throw new SQLException("Delete from ANSWER failed, no rows affected");
@@ -88,7 +88,7 @@ public class AnswerRepository implements Persistent<Answer>{
     }
 
     @Override
-    public Answer findById(int id) {
+    public Answer findById(long id) {
         return null;
     }
 }
